moved runner's cluster back to networking due to CDK's brokenness
