+++ lib/droneRunner.js
@@ -16,20 +16,6 @@ class DroneRunner extends cdk.Construct {
   constructor(scope, props, id = 'DroneRunner') {
     super(scope, id);
 
-    const containerInstance = props.networking.cluster.addCapacity('RunnerCapacity', {
-      instanceType: new ec2.InstanceType(props.parameters.RunnerInstanceType.valueAsString),
-      maxCapacity: props.parameters.RunnerMaxCapacity.valueAsNumber,
-    });
-
-    containerInstance.scaleOnCpuUtilization('RunnerCPUScaling', {
-      targetUtilizationPercent: 90,
-      scaleInCooldown: cdk.Duration.minutes(1),
-      scaleOutCooldown: cdk.Duration.minutes(1),
-    });
-
-    /** @private */
-    this._containerInstance = containerInstance;
-
     const taskDefinition = new ecs.Ec2TaskDefinition(this, 'RunnerTaskDef');
     // Mount docker socket inside runner containers
     taskDefinition.addVolume({
@@ -66,10 +52,6 @@ class DroneRunner extends cdk.Construct {
       assignPublicIp: false,
       taskDefinition,
     });
-  }
-
-  get containerInstance() {
-    return this._containerInstance;
   }
 }
