+++ lib/efsCache.js
@@ -15,7 +15,7 @@ const { Networking } = require('./networking');
 const { DroneRunner } = require('./droneRunner');
 
 class EFSCache extends cdk.Construct {
-  /** @param {{networking:Networking=, parameters:Parameters=, droneRunner:DroneRunner=}} props */
+  /** @param {{networking:Networking=, parameters:Parameters=}} props */
   constructor(scope, props, id = 'EFSCache') {
     super(scope, id);
 
@@ -76,7 +76,7 @@ class EFSCache extends cdk.Construct {
     const mountOpts = 'nfsvers=4.1,rsize=1048576,wsize=1048576,hard,timeo=600,retrans=2,noresvport';
     const mountDest = '/mnt/efs';
     // Inject EFS into Runner ECS Container Instances
-    props.droneRunner.containerInstance.addUserData([
+    props.networking.containerInstance.addUserData([
       'yum update -y',
       'yum install -y amazon-efs-utils',
       `mkdir -p ${mountDest}`,
